crop-recommendation-ai-main/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
numpy
scikit-learn
requests
python-multipart
pydantic
httpx
python-dotenv
